+++ Source/Editors/Library/Lists.cs
@@ -210,7 +210,7 @@ class Lists
             public byte Behaviour;
             public byte SpawnTime;
             public byte Sight;
-            public byte Experience;
+            public int Experience;
             public short[] Vital;
             public short[] Attribute;
             public List<NPC_Drop> Drop;
@@ -235,7 +235,7 @@ class Lists
             public short Req_Level;
             public byte Req_Class;
             // Poção
-            public short Potion_Experience;
+            public int Potion_Experience;
             public short[] Potion_Vital;
             // Equipamento
             public byte Equip_Type;
